Fix transformation ordering to match original VensimExprTranslator
Move power operator (^→**) from OperatorTransformationStage to
SubscriptAndCleanupStage to preserve original position after function
translations. Move GET XLS/DIRECT functions and unsupported function
check from FunctionTranslationStage to SubscriptAndCleanupStage to
preserve original position after lookup call rewriting.
